refactor: '/'에 대응되는 페이지 변경
- 기존 /main의 페이지를 /로 옮겼습니다.
