+++ src/router.ts
@@ -8,7 +8,6 @@ export type Path =
   | `/home`
   | `/login`
   | `/login/:provider`
-  | `/main`
   | `/select-univ`
   | `/signup`;
